Update wheelie turn macro
